move error functions to validation
