+++ tests/models/venue_test.py
@@ -1,7 +1,6 @@
 import pytest
 
 from models import ApiErrors
-from models.venue import TooManyVirtualVenuesException
 from repository import repository
 from tests.conftest import clean_database
 from tests.model_creators.generic_creators import create_offerer, create_venue, create_bank_information
@@ -11,7 +10,8 @@ from tests.model_creators.specific_creators import create_offer_with_thing_produ
 @clean_database
 def test_offerer_cannot_have_address_and_isVirtual(app):
     # Given
-    offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000', name='Test offerer')
+    offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000',
+                             name='Test offerer')
     repository.save(offerer)
 
     venue = create_venue(offerer, name='Venue_name', booking_email='booking@email.com', is_virtual=True, siret=None)
@@ -25,7 +25,8 @@ def test_offerer_cannot_have_address_and_isVirtual(app):
 @clean_database
 def test_offerer_not_isVirtual_and_has_siret_can_have_null_address(app):
     # Given
-    offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000', name='Test offerer')
+    offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000',
+                             name='Test offerer')
     repository.save(offerer)
 
     venue = create_venue(offerer, siret='12345678912345', name='Venue_name', booking_email='booking@email.com',
@@ -43,7 +44,8 @@ def test_offerer_not_isVirtual_and_has_siret_can_have_null_address(app):
 @clean_database
 def test_offerer_not_isVirtual_and_has_siret_cannot_have_null_postal_code_nor_city_nor_departement_code(app):
     # Given
-    offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000', name='Test offerer')
+    offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000',
+                             name='Test offerer')
     repository.save(offerer)
 
     venue = create_venue(offerer, siret='12345678912345', name='Venue_name', booking_email='booking@email.com',
@@ -59,7 +61,8 @@ def test_offerer_not_isVirtual_and_has_siret_cannot_have_null_postal_code_nor_ci
 def test_offerer_not_isVirtual_and_has_no_siret_cannot_have_null_address_nor_postal_code_nor_city_nor_departement_code(
         app):
     # Given
-    offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000', name='Test offerer')
+    offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000',
+                             name='Test offerer')
     repository.save(offerer)
 
     venue = create_venue(offerer, siret=None, name='Venue_name', booking_email='booking@email.com', address=None,
@@ -73,7 +76,8 @@ def test_offerer_not_isVirtual_and_has_no_siret_cannot_have_null_address_nor_pos
 @clean_database
 def test_offerer_not_isVirtual_and_has_no_siret_and_has_address_and_postal_code_and_city_and_departement_code(app):
     # Given
-    offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000', name='Test offerer')
+    offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000',
+                             name='Test offerer')
     repository.save(offerer)
 
     venue = create_venue(offerer, siret=None, comment="fake comment", name='Venue_name',
@@ -93,36 +97,52 @@ def test_offerer_not_isVirtual_and_has_no_siret_and_has_address_and_postal_code_
 @clean_database
 def test_offerer_cannot_create_a_second_virtual_venue(app):
     # Given
-    offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000', name='Test offerer')
+    offerer = create_offerer(siren='123456789',
+                             address='1 rue Test',
+                             city='Test city',
+                             postal_code='93000',
+                             name='Test offerer')
     repository.save(offerer)
-
-    venue = create_venue(offerer, name='Venue_name', booking_email='booking@email.com', address=None, postal_code=None,
-                         city=None, departement_code=None, is_virtual=True, siret=None)
+    venue = create_venue(offerer,
+                         name='Venue_name',
+                         booking_email='booking@email.com',
+                         address=None,
+                         postal_code=None,
+                         city=None,
+                         departement_code=None,
+                         is_virtual=True,
+                         siret=None)
     repository.save(venue)
-
-    new_venue = create_venue(offerer, name='Venue_name', booking_email='booking@email.com', address=None,
-                             postal_code=None, city=None, departement_code=None, is_virtual=True, siret=None)
+    new_venue = create_venue(offerer,
+                             name='Venue_name',
+                             booking_email='booking@email.com',
+                             address=None,
+                             postal_code=None,
+                             city=None,
+                             departement_code=None,
+                             is_virtual=True,
+                             siret=None)
 
     # When
-    with pytest.raises(TooManyVirtualVenuesException):
+    with pytest.raises(ApiErrors) as errors:
         repository.save(new_venue)
+
+    # Then
+    assert errors.value.errors['isVirtual'] == ['Un lieu pour les offres numériques existe déjà pour cette structure']
 
 
 @clean_database
 def test_offerer_cannot_update_a_second_venue_to_be_virtual(app):
     # Given
     siren = '132547698'
-    offerer = create_offerer(siren=siren, address='1 rue Test', city='Test city', postal_code='93000', name='Test offerer')
+    offerer = create_offerer(siren=siren, address='1 rue Test', city='Test city', postal_code='93000',
+                             name='Test offerer')
     repository.save(offerer)
-
     venue = create_venue(offerer, address=None, postal_code=None, city=None, departement_code=None, is_virtual=True,
                          siret=None)
     repository.save(venue)
-
     new_venue = create_venue(offerer, is_virtual=False, siret=siren + '98765')
     repository.save(new_venue)
-
-    # When
     new_venue.isVirtual = True
     new_venue.postalCode = None
     new_venue.address = None
@@ -130,9 +150,12 @@ def test_offerer_cannot_update_a_second_venue_to_be_virtual(app):
     new_venue.departementCode = None
     new_venue.siret = None
 
-    # Then
-    with pytest.raises(TooManyVirtualVenuesException):
+    # When
+    with pytest.raises(ApiErrors) as errors:
         repository.save(new_venue)
+
+    # Then
+    assert errors.value.errors['isVirtual'] == ['Un lieu pour les offres numériques existe déjà pour cette structure']
 
 
 @clean_database
@@ -215,12 +238,15 @@ def test_nOffers(app):
     # then
     assert n_offers == 4
 
+
 class DepartementCodeTest:
     @clean_database
     def test_venue_in_overseas_departemenet_has_a_three_digit_departement_code(self, app):
         # Given
-        offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000', name='Test offerer')
-        venue = create_venue(offerer, siret='12345678912345', name='Venue_name', booking_email='booking@email.com', address='1 test address', postal_code='97300', city='Cayenne', is_virtual=False)
+        offerer = create_offerer(siren='123456789', address='1 rue Test', city='Test city', postal_code='93000',
+                                 name='Test offerer')
+        venue = create_venue(offerer, siret='12345678912345', name='Venue_name', booking_email='booking@email.com',
+                             address='1 test address', postal_code='97300', city='Cayenne', is_virtual=False)
         repository.save(venue)
 
         # When
@@ -228,6 +254,7 @@ class DepartementCodeTest:
 
         # Then
         assert departementCode == '973'
+
 
 class VenueBankInformationTest:
     @clean_database
