+++ tests/validation/headers_test.py
@@ -1,9 +1,9 @@
 from unittest.mock import patch
 
-from validation.headers import check_origin_header_validity
+from validation.routes.headers import check_origin_header_validity
 
 
-@patch('validation.headers.ENV', 'development')
+@patch('validation.routes.headers.ENV', 'development')
 def test_is_valid_header_when_is_dev_and_header_is_local_host_for_normal_endpoint():
     # Given
     header_origin = 'http://localhost:3000'
@@ -16,7 +16,7 @@ def test_is_valid_header_when_is_dev_and_header_is_local_host_for_normal_endpoin
     assert is_valid_header
 
 
-@patch('validation.headers.ENV', 'testing')
+@patch('validation.routes.headers.ENV', 'testing')
 def test_is_valid_header_when_is_testing_and_header_is_local_host_for_normal_endpoint():
     # Given
     header_origin = 'http://localhost:3000'
@@ -29,8 +29,8 @@ def test_is_valid_header_when_is_testing_and_header_is_local_host_for_normal_end
     assert is_valid_header
 
 
-@patch('validation.headers.ENV', 'staging')
-@patch('validation.headers.API_URL', 'https://backend.passculture-staging.beta.gouv.fr')
+@patch('validation.routes.headers.ENV', 'staging')
+@patch('validation.routes.headers.API_URL', 'https://backend.passculture-staging.beta.gouv.fr')
 def test_is_not_valid_header_when_is_staging_and_header_is_local_host_for_normal_endpoint():
     # Given
     header_origin = 'http://localhost:3000'
@@ -43,8 +43,8 @@ def test_is_not_valid_header_when_is_staging_and_header_is_local_host_for_normal
     assert not is_valid_header
 
 
-@patch('validation.headers.ENV', 'staging')
-@patch('validation.headers.API_URL', 'https://backend.passculture-staging.beta.gouv.fr')
+@patch('validation.routes.headers.ENV', 'staging')
+@patch('validation.routes.headers.API_URL', 'https://backend.passculture-staging.beta.gouv.fr')
 def test_is_valid_header_when_header_not_in_whitelist_for_exception_endpoint():
     # Given
     header_origin = 'http://random.url.com'
@@ -57,8 +57,8 @@ def test_is_valid_header_when_header_not_in_whitelist_for_exception_endpoint():
     assert is_valid_header
 
 
-@patch('validation.headers.ENV', 'staging')
-@patch('validation.headers.API_URL', 'https://backend.passculture-staging.beta.gouv.fr')
+@patch('validation.routes.headers.ENV', 'staging')
+@patch('validation.routes.headers.API_URL', 'https://backend.passculture-staging.beta.gouv.fr')
 def test_is_valid_header_when_is_staging_and_header_is_pro_passculture_staging_for_normal_endpoint():
     # Given
     header_origin = 'http://pro.passculture-staging.beta.gouv.fr'
@@ -71,8 +71,8 @@ def test_is_valid_header_when_is_staging_and_header_is_pro_passculture_staging_f
     assert is_valid_header
 
 
-@patch('validation.headers.ENV', 'staging')
-@patch('validation.headers.API_URL', 'https://backend.passculture-staging.beta.gouv.fr')
+@patch('validation.routes.headers.ENV', 'staging')
+@patch('validation.routes.headers.API_URL', 'https://backend.passculture-staging.beta.gouv.fr')
 def test_is_valid_header_when_is_staging_and_header_is_app_passculture_staging_for_normal_endpoint():
     # Given
     header_origin = 'http://app.passculture-staging.beta.gouv.fr'
@@ -85,8 +85,8 @@ def test_is_valid_header_when_is_staging_and_header_is_app_passculture_staging_f
     assert is_valid_header
 
 
-@patch('validation.headers.ENV', 'production')
-@patch('validation.headers.API_URL', 'https://backend.passculture.beta.gouv.fr')
+@patch('validation.routes.headers.ENV', 'production')
+@patch('validation.routes.headers.API_URL', 'https://backend.passculture.beta.gouv.fr')
 def test_is_not_valid_header_when_not_staging_not_dev_and_header_is_app_passculture_staging_for_normal_endpoint():
     # Given
     header_origin = 'http://app.passculture-staging.beta.gouv.fr'
@@ -99,8 +99,8 @@ def test_is_not_valid_header_when_not_staging_not_dev_and_header_is_app_passcult
     assert not is_valid_header
 
 
-@patch('validation.headers.ENV', 'production')
-@patch('validation.headers.API_URL', 'https://backend.passculture.beta.gouv.fr')
+@patch('validation.routes.headers.ENV', 'production')
+@patch('validation.routes.headers.API_URL', 'https://backend.passculture.beta.gouv.fr')
 def test_is_valid_header_when_not_staging_not_dev_for_exception_endpoint_and_exception_endpoint():
     # Given
     header_origin = 'http://random.url.fr'
@@ -113,8 +113,8 @@ def test_is_valid_header_when_not_staging_not_dev_for_exception_endpoint_and_exc
     assert is_valid_header
 
 
-@patch('validation.headers.ENV', 'production')
-@patch('validation.headers.API_URL', 'https://backend.passculture.beta.gouv.fr')
+@patch('validation.routes.headers.ENV', 'production')
+@patch('validation.routes.headers.API_URL', 'https://backend.passculture.beta.gouv.fr')
 def test_is_valid_header_when_not_staging_not_dev_and_header_is_app_passculture_for_normal_endpoint():
     # Given
     header_origin = 'http://app.passculture.beta.gouv.fr'
@@ -127,8 +127,8 @@ def test_is_valid_header_when_not_staging_not_dev_and_header_is_app_passculture_
     assert is_valid_header
 
 
-@patch('validation.headers.ENV', 'production')
-@patch('validation.headers.API_URL', 'https://backend.passculture.beta.gouv.fr')
+@patch('validation.routes.headers.ENV', 'production')
+@patch('validation.routes.headers.API_URL', 'https://backend.passculture.beta.gouv.fr')
 def test_not_valid_header_when_not_staging_not_dev_and_header_is_pro_passculture_for_normal_endpoint():
     # Given
     header_origin = 'http://pro.passculture.beta.gouv.fr'
@@ -141,8 +141,8 @@ def test_not_valid_header_when_not_staging_not_dev_and_header_is_pro_passculture
     assert is_valid_header
 
 
-@patch('validation.headers.ENV', 'production')
-@patch('validation.headers.API_URL', 'https://backend.passculture.beta.gouv.fr')
+@patch('validation.routes.headers.ENV', 'production')
+@patch('validation.routes.headers.API_URL', 'https://backend.passculture.beta.gouv.fr')
 def test_is_valid_header_when_not_staging_not_dev_and_header_is_pro_passculture_for_exception_endpoint():
     # Given
     header_origin = 'http://random.url.frv'
@@ -155,8 +155,8 @@ def test_is_valid_header_when_not_staging_not_dev_and_header_is_pro_passculture_
     assert is_valid_header
 
 
-@patch('validation.headers.ENV', 'production')
-@patch('validation.headers.API_URL', 'https://backend.passculture.beta.gouv.fr')
+@patch('validation.routes.headers.ENV', 'production')
+@patch('validation.routes.headers.API_URL', 'https://backend.passculture.beta.gouv.fr')
 def test_any_origin_header_is_valid_on_endpoint_validate_venue():
     # Given
     header_origin = 'http://random.url.fr'
@@ -169,7 +169,7 @@ def test_any_origin_header_is_valid_on_endpoint_validate_venue():
     assert is_valid_header
 
 
-@patch('validation.headers.ENV', 'testing')
+@patch('validation.routes.headers.ENV', 'testing')
 def test_is_valid_header_when_url_is_authorized_in_subdomain():
     # Given
     header_origin = 'https://poc.passculture.app'
@@ -182,7 +182,7 @@ def test_is_valid_header_when_url_is_authorized_in_subdomain():
     assert is_valid_header
 
 
-@patch('validation.headers.ENV', 'testing')
+@patch('validation.routes.headers.ENV', 'testing')
 def test_is_invalid_header_when_url_is_not_known():
     # Given
     header_origin = 'https://poc.fr'
@@ -195,7 +195,7 @@ def test_is_invalid_header_when_url_is_not_known():
     assert not is_valid_header
 
 
-@patch('validation.headers.ENV', 'testing')
+@patch('validation.routes.headers.ENV', 'testing')
 def test_is_invalid_header_when_malicious_url_is_used():
     # Given
     header_origin = 'https://my.malicious.website.com?origin=poc.passculture.app'
@@ -208,7 +208,7 @@ def test_is_invalid_header_when_malicious_url_is_used():
     assert not is_valid_header
 
 
-@patch('validation.headers.ENV', 'testing')
+@patch('validation.routes.headers.ENV', 'testing')
 def test_is_valid_header_when_url_has_dash_is_authorized_in_subdomain():
     # Given
     header_origin = 'https://my-poc.passculture.app'
