+++ tests/validation/mediations_test.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import pytest
 
 from models import ApiErrors
-from validation.mediations import check_thumb_quality
+from validation.routes.mediations import check_thumb_quality
 
 MODULE_PATH = Path(os.path.dirname(os.path.realpath(__file__)))
