+++ validation/routes/offers.py
