+++ validation/routes/stocks.py
